+++ Components/FragmentTest.php
@@ -141,7 +141,7 @@ final class FragmentTest extends TestCase
             ],
             'PSR-7 URI object with empty string fragment' => [
                 'uri' => Http::new('http://example.com#'),
-                'expected' => '',
+                'expected' => null,
             ],
             'League URI object' => [
                 'uri' => Uri::new('http://example.com#foobar'),